feat(UIKIT-936,icons): Заменена иконка SuccessOutlineSm и перенесена в custom icons
